Consolidate asynchronous serial header/source file pairs
Resolves #1954 (Consolidate asynchronous serial header/source file
pairs).
