+++ include/picolibrary/testing/interactive/asynchronous_serial.h
@@ -23,7 +23,10 @@
 #ifndef PICOLIBRARY_TESTING_INTERACTIVE_ASYNCHRONOUS_SERIAL_H
 #define PICOLIBRARY_TESTING_INTERACTIVE_ASYNCHRONOUS_SERIAL_H
 
+#include <utility>
+
 #include "picolibrary/rom.h"
+#include "picolibrary/testing/interactive/stream.h"
 
 /**
  * \brief Asynchronous serial interactive testing facilities.
@@ -46,6 +49,24 @@ void hello_world( Transmitter transmitter ) noexcept
           auto const character = *string++; ) {
         transmitter.transmit( character );
     } // for
+}
+
+/**
+ * \brief Asynchronous serial reliable output stream hello world interactive test helper.
+ *
+ * \tparam Reliable_Output_Stream The type of reliable output stream to use.
+ * \tparam Transmitter The type of transmitter to use.
+ *
+ * \param[in] transmitter The transmitter to use.
+ */
+template<template<typename> typename Reliable_Output_Stream, typename Transmitter>
+void hello_world( Transmitter transmitter ) noexcept
+{
+    transmitter.initialize();
+
+    auto stream = Reliable_Output_Stream{ std::move( transmitter ) };
+
+    ::picolibrary::Testing::Interactive::hello_world( stream );
 }
 
 } // namespace picolibrary::Testing::Interactive::Asynchronous_Serial
