+++ source/CMakeLists.txt
@@ -28,7 +28,6 @@ set(
     "picolibrary/algorithm.cc"
     "picolibrary/array.cc"
     "picolibrary/asynchronous_serial.cc"
-    "picolibrary/asynchronous_serial/stream.cc"
     "picolibrary/bit_manipulation.cc"
     "picolibrary/circular_buffer.cc"
     "picolibrary/crc.cc"
@@ -120,7 +119,6 @@ if( ${PICOLIBRARY_ENABLE_INTERACTIVE_TESTING} )
         "picolibrary/testing/interactive.cc"
         "picolibrary/testing/interactive/adc.cc"
         "picolibrary/testing/interactive/asynchronous_serial.cc"
-        "picolibrary/testing/interactive/asynchronous_serial/stream.cc"
         "picolibrary/testing/interactive/gpio.cc"
         "picolibrary/testing/interactive/i2c.cc"
         "picolibrary/testing/interactive/ip.cc"
