+++ test/automated/picolibrary/asynchronous_serial/reliable_unbuffered_output_stream/main.cc
@@ -25,7 +25,7 @@
 
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
-#include "picolibrary/asynchronous_serial/stream.h"
+#include "picolibrary/asynchronous_serial.h"
 #include "picolibrary/testing/automated/asynchronous_serial.h"
 
 namespace {
